simulatedMeter/requirements.txt
paho-mqtt>=1.6.1
requests>=2.28.0
PyYAML>=6.0
python-dotenv>=1.0.0 